+++ akshare/__init__.py
@@ -2524,9 +2524,10 @@ amac_manager_cancelled_info # 中国证券投资基金业协会-信息公示-诚
 1.11.8 fix: fix news_economic_baidu interface
 1.11.9 fix: fix currency_boc_safe interface
 1.11.10 fix: fix bond_new_composite_index_cbond interface
+1.11.11 fix: fix spot_hist_sge interface
 """
 
-__version__ = "1.11.10"
+__version__ = "1.11.11"
 __author__ = "AKFamily"
 
 import sys