+++ akshare/spot/spot_sge.py
@@ -1,7 +1,7 @@
 # -*- coding:utf-8 -*-
 # !/usr/bin/env python
 """
-Date: 2022/11/17 22:00
+Date: 2023/9/19 8:00
 Desc: 上海黄金交易所-数据资讯-行情走势
 https://www.sge.com.cn/sjzx/mrhq
 上海黄金交易所-数据资讯-上海金基准价-历史数据
@@ -18,37 +18,29 @@ def spot_symbol_table_sge() -> pd.DataFrame:
     :return: 品种表
     :rtype: pandas.DataFrame
     """
-    url = "https://www.sge.com.cn/sjzx/mrhq_data"
-    payload = {"date": "2022-11-17"}
-    headers = {
-        "Accept": "text/html, */*; q=0.01",
-        "Accept-Encoding": "gzip, deflate, br",
-        "Accept-Language": "zh-CN,zh;q=0.9,en;q=0.8",
-        "Cache-Control": "no-cache",
-        "Connection": "keep-alive",
-        "Content-Length": "15",
-        "Content-Type": "multipart/form-data; boundary=---011000010111000001101001",
-        "Host": "www.sge.com.cn",
-        "Origin": "https://www.sge.com.cn",
-        "Pragma": "no-cache",
-        "Referer": "https://www.sge.com.cn/sjzx/mrhq",
-        "sec-ch-ua": '"Google Chrome";v="107", "Chromium";v="107", "Not=A?Brand";v="24"',
-        "sec-ch-ua-mobile": "?0",
-        "sec-ch-ua-platform": '"Windows"',
-        "Sec-Fetch-Dest": "empty",
-        "Sec-Fetch-Mode": "cors",
-        "Sec-Fetch-Site": "same-origin",
-        "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/107.0.0.0 Safari/537.36",
-        "X-Requested-With": "XMLHttpRequest",
-    }
-    r = requests.post(url, data=payload, headers=headers)
-    data_text = r.text
-    re_html = "<" + data_text.strip().strip("<!--").strip("-->") + ">"
-    temp_df = pd.read_html(re_html)[0]
+    temp_list = [
+        "Au99.99",
+        "Au99.95",
+        "Au100g",
+        "Pt99.95",
+        "Ag(T+D)",
+        "Au(T+D)",
+        "mAu(T+D)",
+        "Au(T+N1)",
+        "Au(T+N2)",
+        "Ag99.99",
+        "iAu99.99",
+        "Au99.5",
+        "iAu100g",
+        "iAu99.5",
+        "PGC30g",
+        "NYAuTN06",
+        "NYAuTN12",
+    ]
+    temp_df = pd.DataFrame(temp_list)
     temp_df.reset_index(inplace=True)
-    temp_df = temp_df.iloc[:, :2]
     temp_df.columns = ["序号", "品种"]
-    temp_df['序号'] = temp_df['序号'] + 1
+    temp_df["序号"] = temp_df["序号"] + 1
     return temp_df
 
 
@@ -95,11 +87,11 @@ def spot_hist_sge(symbol: str = "Au99.99") -> pd.DataFrame:
         "low",
     ]
 
-    temp_df["date"] = pd.to_datetime(temp_df["date"]).dt.date
-    temp_df["open"] = pd.to_numeric(temp_df["open"])
-    temp_df["close"] = pd.to_numeric(temp_df["close"])
-    temp_df["high"] = pd.to_numeric(temp_df["high"])
-    temp_df["low"] = pd.to_numeric(temp_df["low"])
+    temp_df["date"] = pd.to_datetime(temp_df["date"], errors="coerce").dt.date
+    temp_df["open"] = pd.to_numeric(temp_df["open"], errors="coerce")
+    temp_df["close"] = pd.to_numeric(temp_df["close"], errors="coerce")
+    temp_df["high"] = pd.to_numeric(temp_df["high"], errors="coerce")
+    temp_df["low"] = pd.to_numeric(temp_df["low"], errors="coerce")
     return temp_df
 
 
@@ -128,10 +120,12 @@ def spot_golden_benchmark_sge() -> pd.DataFrame:
         "交易时间",
         "早盘价",
     ]
-    temp_zp_df["交易时间"] = pd.to_datetime(temp_zp_df["交易时间"], unit="ms").dt.date
+    temp_zp_df["交易时间"] = pd.to_datetime(
+        temp_zp_df["交易时间"], unit="ms", errors="coerce"
+    ).dt.date
     temp_df["早盘价"] = temp_zp_df["早盘价"]
-    temp_df["晚盘价"] = pd.to_numeric(temp_df["晚盘价"])
-    temp_df["早盘价"] = pd.to_numeric(temp_df["早盘价"])
+    temp_df["晚盘价"] = pd.to_numeric(temp_df["晚盘价"], errors="coerce")
+    temp_df["早盘价"] = pd.to_numeric(temp_df["早盘价"], errors="coerce")
     return temp_df
 
 
@@ -160,10 +154,12 @@ def spot_silver_benchmark_sge() -> pd.DataFrame:
         "交易时间",
         "早盘价",
     ]
-    temp_zp_df["交易时间"] = pd.to_datetime(temp_zp_df["交易时间"], unit="ms").dt.date
+    temp_zp_df["交易时间"] = pd.to_datetime(
+        temp_zp_df["交易时间"], unit="ms", errors="coerce"
+    ).dt.date
     temp_df["早盘价"] = temp_zp_df["早盘价"]
-    temp_df["晚盘价"] = pd.to_numeric(temp_df["晚盘价"])
-    temp_df["早盘价"] = pd.to_numeric(temp_df["早盘价"])
+    temp_df["晚盘价"] = pd.to_numeric(temp_df["晚盘价"], errors="coerce")
+    temp_df["早盘价"] = pd.to_numeric(temp_df["早盘价"], errors="coerce")
     return temp_df
 
 
@@ -176,6 +172,9 @@ if __name__ == "__main__":
 
     spot_golden_benchmark_sge_df = spot_golden_benchmark_sge()
     print(spot_golden_benchmark_sge_df)
+
+    spot_silver_benchmark_sge_df = spot_silver_benchmark_sge()
+    print(spot_silver_benchmark_sge_df)
 
     for spot in spot_symbol_table_sge_df["品种"].tolist():
         spot_hist_sge_df = spot_hist_sge(symbol=spot)